+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## SAS Portal Framework for SAS Viya v1.1.0
 
 - Add: Data Product Registry object
+- Add: Data Product Marketplace object
 - Add: documentation for the Run Custom Code object
 - Add: docuemntation for the Data Product Registry object
+- Add: docuemntation for the Data Product Marketplace object
+- Change: Provided updates to the global.css file to support special cases from the Data Product Marketplace object
 
 ## SAS Portal Framework for SAS Viya v1.0.4